pyproject.toml
[tool.poetry]
name = "caricature-generator"
version = "0.1.0"
description = "Pipeline for generating tasteful caricatures from images."
authors = ["Shubham <shubham@example.com>"]
readme = "README.md"
packages = [{ include = "caricature_generator", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.2.0"
torchvision = "^0.17.0"
diffusers = "^0.27.2"
transformers = "^4.39.0"
opencv-python = "^4.9.0"
mediapipe = "^0.10.11"
numpy = "^1.26.0"
pillow = "^10.2.0"
pydantic = "^2.6.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
loguru = "^0.7.2"
typer = "^0.9.0"
tqdm = "^4.66.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^4.1.0"
ruff = "^0.3.4"
black = "^24.2.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
caricature-pipeline = "caricature_generator.cli:app"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "N", "Q"]
extend-ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true

